+++ src/index.js
@@ -1,7 +1,6 @@
 import './style.css';
-import { Loop } from './modules/API.js';
+import { GetDataFromAPI } from './modules/API.js';
 
 window.onload = async () => {
-  Loop();
-  console.log('HI');
+  GetDataFromAPI();
 };
