--- src/modules/popup.js
@@ -1,8 +1,0 @@
-import renderModal from './renderModal.js';
-import { GetComments } from './API.js';
-
-const openModal = (movie) => {
-  renderModal(movie);
-};
-
-export default openModal;
