+++ src/modules/renderModal.js
@@ -1,10 +1,5 @@
-/* eslint-disable no-unused-vars */
-import { GetComments } from './API';
-
 const renderModal = async (movie) => {
   const modalContainer = document.getElementById('modal-container');
-
-  const commentArray = await GetComments(movie.show.id);
 
   modalContainer.innerHTML = `
       <div class="modal-header">
@@ -20,20 +15,6 @@ const renderModal = async (movie) => {
       <ul id="comment-list"><ul>
       `;
 
-  commentArray.forEach((item) => {
-    const li = document.createElement('li');
-    const name = document.createElement('span');
-
-    name.innerHTML = `${item.username} : ${item.comment}`;
-    li.appendChild(name);
-
-    document.getElementById('comment-list').appendChild(li);
-  });
-
-  const li = document.createElement('li');
-
-  const list = document.getElementById('comment-list');
-  console.log(list);
   document.getElementById('close').addEventListener('click', () => {
     document.getElementById('modal-container').classList.remove('active');
     document.getElementById('overlay').classList.remove('active');
